requirements.txt
dj-database-url==0.5.0
Django==2.1.5
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
Pillow==5.2.0